+++ src/app/router/router.tsx
@@ -2,6 +2,8 @@ import { createBrowserRouter } from 'react-router-dom'
 import App from '../../App'
 import { ROUTES } from '../../utils/constans'
 import { lazy } from 'react'
+import { PrivateRoute } from '../PrivateRoute/PrivateRoute'
+import FavoritesPage from '../../pages/FavoritesPage/FavoritesPage'
 
 const MainPage = lazy(() => import('../../pages/MainPage/MainPage'))
 const SearchPage = lazy(() => import('../../pages/SearchPage/SearchPage'))
@@ -25,6 +27,11 @@ export const router = createBrowserRouter([
       {
         path: `${ROUTES.FILM}/:film`,
         element: <FilmPage />,
+      },
+      {
+        path: '',
+        element: <PrivateRoute />,
+        children: [{ path: 'favorites', element: <FavoritesPage /> }],
       },
       { path: '/signup', element: <SignUpPage /> },
       { path: '/signin', element: <SignInPage /> },
