+++ src/components/ListCards/ListCards.tsx
@@ -1,30 +1,12 @@
-import {
-  Card,
-  CardActionArea,
-  CardActions,
-  CardMedia,
-  Grid,
-  IconButton,
-  Typography,
-} from '@mui/material'
-import FavoriteIcon from '@mui/icons-material/Favorite'
-
-import { useNavigate } from 'react-router-dom'
+import { Grid } from '@mui/material'
 import type { ResultFilms } from '../../app/store/api/transformResponses/types'
+import FilmCard from '../Card/Card'
 
 type PropsListCards = {
   films: ResultFilms[]
 }
 
 export function ListCards({ films }: PropsListCards) {
-  const navigate = useNavigate()
-
-  const handleClick = (id: number | null) => {
-    if (id) {
-      navigate('/film/' + id)
-    }
-  }
-
   return (
     <Grid
       direction='row'
@@ -33,36 +15,7 @@ export function ListCards({ films }: PropsListCards) {
       columnSpacing={4}
     >
       {films.map(film => {
-        return (
-          <Grid key={film.id} width='300px' height='500px' item>
-            <Card>
-              <CardActionArea onClick={() => handleClick(film.id)}>
-                <CardMedia
-                  component='img'
-                  height='400'
-                  image={film.posterUrlPreview}
-                  alt='Paella dish'
-                />
-
-                <Typography
-                  component='p'
-                  sx={{
-                    textDecoration: 'ellipsis',
-                    fontSize: '18px',
-                    padding: '5px',
-                  }}
-                >
-                  {film.nameRu}
-                </Typography>
-              </CardActionArea>
-              <CardActions disableSpacing sx={{ padding: 0 }}>
-                <IconButton aria-label='add to favorites'>
-                  <FavoriteIcon />
-                </IconButton>
-              </CardActions>
-            </Card>
-          </Grid>
-        )
+        return <FilmCard key={film.id} film={film} />
       })}
     </Grid>
   )
