+++ src/pages/FavoritesPage/FavoritesPage.tsx
@@ -1,0 +1,3 @@
+export default function FavoritesPage() {
+  return <h1>favorites</h1>
+}
